+++ src/omni_rewriter/models/__init__.py
@@ -11,6 +11,7 @@ from .lingbot import (
     LingBotProminentElement,
     LingBotVideoDescription,
 )
+from .ltx import LTXRenderMode, LTXRewrite, render_ltx_output
 from .media import MediaReference
 from .observation import (
     DialogueObservation,
@@ -34,7 +35,7 @@ from .seedance import (
     render_seedance_output,
 )
 
-RewriteOutput = BaseRewrite | Ref2VARewrite | ImageRewrite | SeedanceRewrite
+RewriteOutput = BaseRewrite | Ref2VARewrite | ImageRewrite | SeedanceRewrite | LTXRewrite
 
 __all__ = [
     "ALLOWED_RATIOS",
@@ -43,6 +44,8 @@ __all__ = [
     "IMAGE_TASKS",
     "ImagePEProfile",
     "ImageRewrite",
+    "LTXRenderMode",
+    "LTXRewrite",
     "LingBotAction",
     "LingBotCameraInfo",
     "LingBotCaption",
@@ -71,5 +74,6 @@ __all__ = [
     "format_timecode",
     "infer_task",
     "parse_timecode",
+    "render_ltx_output",
     "render_seedance_output",
 ]